WaniKani_AI/requirements.txt
openai
requests
pyperclip
pillow
gtts
python-dotenv